+++ grails-app/conf/spring/resources.groovy
@@ -1,8 +1,5 @@
 import org.springframework.jdbc.datasource.DriverManagerDataSource
-import humulus.Environment
-import humulus.SwitchableDataSource
 import humulus.CinnamonPasswordEncoder
-import humulus.RepositoryLoginFilter
 
 import cinnamon.CinnamonAuthenticationDetailsSource
 import cinnamon.CinnamonPreAuthUserDetailsService
@@ -15,38 +12,6 @@ import org.springframework.security.authentication.dao.DaoAuthenticationProvider
 // Place your Spring DSL code here
 beans = {
     
-    parentDataSource(DriverManagerDataSource) { bean ->
-        bean.'abstract' = true;
-        username = "sa"
-    }
-
-    Environment.list().each { env ->
-        "${env.prefix}DataSource"(DriverManagerDataSource) { bean ->
-            bean.parent = parentDataSource
-            bean.scope = 'prototype'
-            url = env.dbConnectionUrl
-            log.debug("url = '$url'")
-            driverClassName = env.driverClassName
-            if (env.username) {
-                username = env.username
-            }
-            if (env.password) {
-                password = env.password
-            }
-            
-        }
-    }
-
-    def dataSources = [:]
-    Environment.list().each {env ->
-        dataSources[env.id] = ref(env.prefix + 'DataSource')
-        log.debug("dataSource: ${dataSources[env.id]}")
-    }
-
-    dataSource(SwitchableDataSource) {
-        targetDataSources = dataSources
-    }
-
     userDetailsService(CinnamonUserDetailsService){
         // looks like this service is not injected automatically:
         repositoryService = ref('repositoryService')
@@ -78,11 +43,6 @@ beans = {
         providers = ref('authenticationProvider')
     }  
     
-    repositoryLoginFilter(RepositoryLoginFilter){
-        authenticationManager = ref('authenticationManager')
-        grailsApplication = ref('grailsApplication')
-    }
-
     cinnamonAuthenticationDetailsSource(CinnamonAuthenticationDetailsSource){
         grailsApplication = ref('grailsApplication')
     }