+++ grails-app/services/cinnamon/CinnamonPreAuthUserDetailsService.groovy
@@ -1,7 +1,5 @@
 package cinnamon
 
-import humulus.Environment
-import humulus.EnvironmentHolder
 import cinnamon.global.ConfThreadLocal
 import org.springframework.security.core.Authentication
 import org.springframework.security.core.userdetails.AuthenticationUserDetailsService
@@ -21,15 +19,6 @@ class CinnamonPreAuthUserDetailsService implements AuthenticationUserDetailsServ
         }
         
         def repositoryName = ticket.split('@')[1]
-        def env = Environment.list().find {it.dbName == repositoryName}
-        if (!env) {
-            throw new PreAuthenticatedCredentialsNotFoundException("Could not find environment '${repositoryName}'")
-        }
-
-        //test connection
-        EnvironmentHolder.setEnvironment(env)
-        def ds = grailsApplication.getMainContext().dataSource
-        ds.getConnection()
         def cmnSession = Session.findByTicket(ticket)
         if (!cmnSession) {
             throw new PreAuthenticatedCredentialsNotFoundException("Session not found for ticket '${ticket}'")